chore(angular): add e2e tests
