+++ packages/angular/plugins/angular-vite-compiler/angular-vite-plugin.ts
