+++ packages/angular/plugins/angular-vite-compiler/benchmark.ts
