+++ packages/angular/plugins/angular-vite-compiler/cache.ts
