+++ packages/angular/plugins/angular-vite-compiler/host.ts
