+++ packages/angular/plugins/angular-vite-compiler/index.ts
