+++ packages/angular/plugins/angular-vite-compiler/ngcc_processor.ts
