+++ packages/angular/plugins/angular-vite-compiler/path.ts
