+++ packages/angular/plugins/angular-vite-compiler/plugin-options.ts
