+++ packages/angular/plugins/angular-vite-compiler/resolver.ts
