+++ packages/angular/plugins/angular-vite-compiler/symbol.ts
