+++ packages/angular/plugins/angular-vite-compiler/transformation.ts
