+++ packages/angular/plugins/angular-vite-compiler/transformers/replace_resources.ts
