+++ packages/angular/plugins/angular-vite-compiler/utils/optimizer.ts
