+++ packages/angular/src/index.ts
@@ -1,2 +1,2 @@
-import AngularVite from './plugins/vite';
+import AngularVite from '../plugins/vite';
 export default AngularVite;
